requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.2
langchain-core==1.0.1
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pillow==11.3.0
pydantic==2.12.3
pydantic_core==2.41.4
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==70.2.0
sniffio==1.3.1
sympy==1.13.1
tenacity==9.1.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
xxhash==3.6.0
zstandard==0.25.0
